Added check before cost management removal

+++ modules/ADE/private/Governance/Remove-AzureCostManagementBudget.ps1
@@ -4,10 +4,12 @@ function Remove-AzureCostManagementBudget {
     )
 
     Write-ScriptSection "Removing Azure Cost Management Budget"
-    
     $consumptionBudgetName = $armParameters.adeBudgetName
-    az consumption budget delete --budget-name $consumptionBudgetName
-    Confirm-LastExitCode
+    
+    if (Confirm-AzureResourceExists 'budget' $consumptionBudgetName) {
+        az consumption budget delete --budget-name $consumptionBudgetName
+        Confirm-LastExitCode
+    }
 
     Write-ScriptSection "Finished Removing Azure Cost Management Budget"
 }
